refactor: firebase 유틸함수들을 일관성있게 코드와 디렉토리 구조 수정 (#37)
